This PR:
 - Adds a test that `preprocess` can be canceled by closing the channel.
 - Checks the channel periodically in `preprocess` and aborts early if it is canceled.
 - Uses memory more efficiently.
 - Drops twenty-first's `MerkleTree` as a dependency for the module `Pow`, in favor of a local struct `MTree`. Motivation:
    - Use 50% less memory by operating in-place on pre-allocated memory buffers.
    - Enable periodically checking whether the channel is canceled.
 - Adds a test that the Merkle tree step in `preprocess` can be canceled by closing the channel.

While periodically checking the channel for cancellation induces a net performance degradation, this net degradation is fully canceled by two factors:
 - avoiding unnecessary reallocations using `MaybeUninit` and unsafe code, and
 - Merkle tree construction is (surprisingly) faster.

More detailed, intermediate performance summaries are given in the respective commits. Here is the delta between master and this PR.


| merkle tree height |   master    | this PR |
|:------------------:|:-------------:|:---------------:|
|                 25 |   9 s / 2 GB  |    9 s / 2 GB   |
|                 26 |  19 s / 5 GB  |   26 s / 5 GB   |
|                 27 |  39 s / 10 GB |   38 s / 10 GB  |
|                 28 |  82 s / 21 GB |   78 s / 21 GB  |
|                 29 | 163 s / 42 GB |  159 s / 42 GB  |